# sentinelapi/sentinelshop/openapi.yaml
openapi: 3.0.3
info:
  title: SentinelShop API
  description: E-Commerce target API contract for security audit.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local demo environment
  - url: http://sentinelshop:4000
    description: Docker container environment

paths:
  /auth/login:
    post:
      summary: User login
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  token_type:
                    type: string
                  userId:
                    type: string
                  role:
                    type: string
        '401':
          description: Invalid credentials

  /auth/register:
    post:
      summary: Register new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string

  /auth/logout:
    post:
      summary: User logout
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully

  /users/me:
    get:
      summary: Get authenticated user profile
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - email
                  - role
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '401':
          description: Unauthorized

  /users/{user_id}:
    get:
      summary: Get user profile by ID
      operationId: getUserProfile
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '403':
          description: Forbidden
        '404':
          description: Not found

  /orders:
    get:
      summary: List current user orders
      operationId: listOrders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Create new order
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - item
                - amount
              properties:
                item:
                  type: string
                amount:
                  type: number
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/{order_id}:
    get:
      summary: Get order details by ID
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Cancel or delete order by ID
      operationId: deleteOrder
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order deleted
        '403':
          description: Forbidden
        '404':
          description: Not found

  /admin/users:
    get:
      summary: List all users in system (Admin role required)
      operationId: adminListUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of system users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
                    email:
                      type: string
                    role:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (insufficient permissions)

  /admin/dashboard:
    get:
      summary: Admin dashboard operational metrics (Admin role required)
      operationId: adminDashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Admin dashboard statistics
        '403':
          description: Forbidden

  /invoices:
    get:
      summary: List invoices (Confidential financial records)
      operationId: listInvoices
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    userId:
                      type: integer
                    invoiceNumber:
                      type: string
                    amount:
                      type: number
                    tax:
                      type: number
        '401':
          description: Unauthorized

  /health:
    get:
      summary: Service health check
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: Service healthy

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Order:
      type: object
      required:
        - id
        - userId
        - item
        - amount
        - status
      properties:
        id:
          type: integer
        userId:
          type: integer
        item:
          type: string
        amount:
          type: number
        status:
          type: string